Refactor to use XSS-safe messages in http service.
