+++ neuro_san/service/mcp/processors/mcp_ping_processor.py
@@ -21,7 +21,7 @@ from typing import Any
 from typing import Dict
 
 from neuro_san.service.http.logging.http_logger import HttpLogger
-from neuro_san.service.mcp.util.mcp_request_util import McpRequestUtil
+from neuro_san.service.utils.request_util import RequestUtil
 
 
 class McpPingProcessor:
@@ -44,6 +44,6 @@ class McpPingProcessor:
         _ = metadata
         return {
             "jsonrpc": "2.0",
-            "id": McpRequestUtil.safe_request_id(request_id),
+            "id": RequestUtil.safe_request_id(request_id),
             "result": {}
         }
